backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.37.28
botocore==1.37.28
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
databases==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
Mako==1.3.9
MarkupSafe==3.0.2
mysqlclient
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyMySQL
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
